requirements.txt
africastalking==1.1.9
autopep8==1.4.4
black==19.10b0
certifi==2019.9.11
chardet==3.0.4
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.1
django-extensions==2.2.5
django-heroku==0.3.1
django-rest-passwordreset==1.1.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.0
gunicorn==20.0.4
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
packaging==20.1
Pillow==6.2.1
pre-commit==2.2.0
psycopg2==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
pyparsing==2.4.6
python-decouple==3.3
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.7
ruamel.yaml.clib==0.2.0
schema==0.7.1
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==5.0.1
